requirements.txt
asgiref==3.8.1
Django==5.1.4
django-otp==1.5.4
djangorestframework==3.15.2
pyotp==2.9.0
sqlparse==0.5.3
